Refactorings

+++ src/main/java/transactions/xa/AccountDao.java
@@ -4,19 +4,19 @@ import transactions.jpa.entity.Account;
 
 public interface AccountDao {
 
-	/**
-	 * Save account into database.
-	 * 
-	 * @param account
-	 *            - account to be saved.
-	 */
-	void save(Account account);
+    /**
+     * Save account into database.
+     * 
+     * @param account
+     *            - account to be saved.
+     */
+    void save(Account account);
 
-	/**
-	 * counts all the accounts.
-	 * 
-	 * @return - the number of accounts.
-	 */
-	Integer getAllCount();
+    /**
+     * counts all the accounts.
+     * 
+     * @return - the number of accounts.
+     */
+    long getAllCount();
 
 }
